js-splitting-code-with-babel
